FunctionApp/Data Connectors/requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
azure-monitor-ingestion==1.0.4
azure-core==1.32.0
